Moved admin endpoint to another file into endpoints folder
